base extension pushed
* this commit creates the base version of the extension.
* all components are 'vanilla' versions of the base components.

In comparison with the existing LibrOffice extensions, this commit only
adds proofingtools to Portuguese (Angola) locale.
This completes support for that language and allows the use of pre-AO90
Portuguese tools in a way that is compatible with local regulations and
other programs, e.g. LanguageTool.
All European Portuguese speaker, using the pre-AO90 agreement (me
included) can change the language of the documento to pt-AO while
writting and reviewing and reformat the documento to Portugal Portuguese
when the document is finalized.